feat: partial support for no {} array subscripts
e.g. $commands[nvim]

It works as part of a test condition but not as part of a command
because the concatenation rule keeps getting picked and it splits the
array from it's subscript. I tried mucking around with precedence but
haven't been able to resolve it yet. Added a currently failing test case
(Command with array subscript)