feat: echo $# $#var
This was a really tough one for me to figure out because _concat kept
grabbing all of the input. Finally realized that _concat doesn't consume
any tokes but if it matches, it'll add the other tokens and it'll match
in a lot of cases that are more nuanced in zsh.